new Test

+++ src/main/resources/db/mysql/schema.sql
@@ -1,3 +1,4 @@
+
 CREATE TABLE IF NOT EXISTS vets (
   id INT(4) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
   first_name VARCHAR(30),
@@ -33,7 +34,7 @@ CREATE TABLE IF NOT EXISTS owners (
   address VARCHAR(255),
   city VARCHAR(80),
   telephone VARCHAR(20),
-  age INTEGER(4),
+  age INT(4),
   INDEX(last_name)
 ) engine=InnoDB;
